+++ src/middleware/hideFooter.ts
@@ -1,0 +1,10 @@
+import { Request, Response, NextFunction } from "express";
+
+export default function hideFooter(
+  _req: Request,
+  res: Response,
+  next: NextFunction,
+) {
+  res.locals.hideFooter = true;
+  next();
+}